# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sacoor Task Api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "Development"
        }
    ],
    "paths": {
        "/products/filter": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Filter Json file by name and pvp",
                "operationId": "4b6d6e52bca455e5662837cea9a0e55b",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "title": "name",
                                        "type": "string"
                                    },
                                    "pvp": {
                                        "title": "pvp",
                                        "type": "string"
                                    },
                                    "file": {
                                        "title": "Json file",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "XML response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "invalid file contents",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}